requirements.txt
Django==1.10.5
django-bootstrap3==7.1.0
django-crispy-forms==1.6.1
django-datatable==0.2.1
django-datatable-view==0.9.0b5
django-datatables-view==1.13.2
django-report-builder==3.4.4
django-tables2==1.2.7
